+++ docs/changelog.md
@@ -13,6 +13,7 @@ category: Administration
 
 * refactor: configuration service can be final in template service
 * refactor: improve filtering duplicates during template import
+* test: add tests for upgrade tasks
 
 ### [23.09.2] - 2023-09-17